test: added new mock for transaction, updated deal ticket and withdraw tests
