+++ apps/trading-e2e/src/integration/withdraw.cy.ts
@@ -66,15 +66,7 @@ describe('withdraw', { tags: '@smoke' }, () => {
   });
 
   it('triggers transaction when submitted', () => {
-    cy.mockVegaCommandSync({
-      txHash: 'test-tx-hash',
-      tx: {
-        signature: {
-          value:
-            'd86138bba739bbc1069b3dc975d20b3a1517c2b9bdd401c70eeb1a0ecbc502ec268cf3129824841178b8b506b0b7d650c76644dbd96f524a6cb2158fb7121800',
-        },
-      },
-    });
+    cy.mockVegaWalletTransaction();
     selectAsset(asset1Name);
     cy.getByTestId('BALANCE_AVAILABLE_label').should(
       'contain.text',
