+++ libs/cypress/src/lib/commands/mock-transaction-response.ts
@@ -1,8 +1,5 @@
 import merge from 'lodash/merge';
-import type {
-  V1TransactionResponse,
-  TransactionResponse,
-} from '@vegaprotocol/wallet';
+import type { TransactionResponse } from '@vegaprotocol/wallet';
 import type { PartialDeep } from 'type-fest';
 
 declare global {
@@ -10,7 +7,6 @@ declare global {
   namespace Cypress {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     interface Chainable<Subject> {
-      mockVegaCommandSync(override: PartialDeep<V1TransactionResponse>): void;
       mockVegaWalletTransaction(
         override?: PartialDeep<TransactionResponse>
       ): void;
@@ -18,27 +14,28 @@ declare global {
   }
 }
 
-export function addMockVegaWalletCommands() {
+export function addMockTransactionResponse() {
   Cypress.Commands.add(
-    'mockVegaCommandSync',
-    (override?: PartialDeep<V1TransactionResponse>) => {
+    'mockVegaWalletTransaction',
+    (override?: PartialDeep<TransactionResponse>) => {
       const defaultTransactionResponse = {
-        txHash: 'tx-hash',
+        transactionHash: 'test-tx-hash',
         sentAt: new Date().toISOString(),
         receivedAt: new Date().toISOString(),
-        tx: {
+        transaction: {
           input_data:
             'CPe6vpiqsPqxDBDC1w7KPkoKQGE4Y2M0NjUwMjhiMGY4OTM4YTYzZTEzNDViYzM2ODc3ZWRmODg4MjNmOWU0ZmI4ZDRlN2VkMmFlMzAwNzA3ZTMYASABKAM4Ag==',
           signature: {
             // sig value needs to be 'real' so sigToId function doesn't error out
-            value: 'signature',
+            value:
+              'd86138bba739bbc1069b3dc975d20b3a1517c2b9bdd401c70eeb1a0ecbc502ec268cf3129824841178b8b506b0b7d650c76644dbd96f524a6cb2158fb7121800',
             algo: 'vega/ed25519',
             version: 1,
           },
           From: {
             PubKey: Cypress.env('VEGA_PUBLIC_KEY'),
           },
-          version: 2,
+          version: 3,
           pow: {
             tid: '0CEEC2FDFDB5D68BC0C1E2640440E4AA11E49986CB2929E0F3572E16CB7DC59C',
             nonce: 23525,
@@ -46,24 +43,12 @@ export function addMockVegaWalletCommands() {
         },
       };
 
-      cy.intercept('POST', 'http://localhost:1789/api/v1/command/sync', {
-        body: merge(defaultTransactionResponse, override),
-      }).as('VegaCommandSync');
-    }
-  );
-
-  Cypress.Commands.add(
-    'mockVegaWalletTransaction',
-    (override?: PartialDeep<TransactionResponse>) => {
-      const defaultTransactionResponse = {
-        transactionHash: 'tx-hash',
-        signature: 'signature',
-        sentAt: new Date().toISOString(),
-        receivedAt: new Date().toISOString(),
-      };
-
       cy.intercept('POST', 'http://localhost:1789/api/v2/requests', {
-        body: merge(defaultTransactionResponse, override),
+        body: {
+          jsonrpc: '2.0',
+          result: merge(defaultTransactionResponse, override),
+          id: '1',
+        },
       }).as('VegaWalletTransaction');
     }
   );
